+++ src/mlpack/methods/ann/layer/layer_types.hpp
@@ -36,6 +36,7 @@
 #include <mlpack/methods/ann/layer/multiply_constant.hpp>
 #include <mlpack/methods/ann/layer/max_pooling.hpp>
 #include <mlpack/methods/ann/layer/mean_pooling.hpp>
+#include <mlpack/methods/ann/layer/noisylinear.hpp>
 #include <mlpack/methods/ann/layer/adaptive_max_pooling.hpp>
 #include <mlpack/methods/ann/layer/adaptive_mean_pooling.hpp>
 #include <mlpack/methods/ann/layer/parametric_relu.hpp>
@@ -83,6 +84,10 @@ template<typename InputDataType,
          typename OutputDataType,
          typename RegularizerType>
 class LinearNoBias;
+
+template<typename InputDataType,
+         typename OutputDataType>
+class NoisyLinear;
 
 template<typename InputDataType,
          typename OutputDataType
@@ -193,6 +198,8 @@ template <typename InputDataType,
 class AdaptiveMeanPooling;
 
 using MoreTypes = boost::variant<
+        Glimpse<arma::mat, arma::mat>*,
+        Highway<arma::mat, arma::mat>*,
         Recurrent<arma::mat, arma::mat>*,
         RecurrentAttention<arma::mat, arma::mat>*,
         ReinforceNormal<arma::mat, arma::mat>*,
@@ -238,10 +245,8 @@ using LayerTypes = boost::variant<
     ELU<arma::mat, arma::mat>*,
     FastLSTM<arma::mat, arma::mat>*,
     FlexibleReLU<arma::mat, arma::mat>*,
-    Glimpse<arma::mat, arma::mat>*,
     GRU<arma::mat, arma::mat>*,
     HardTanH<arma::mat, arma::mat>*,
-    Highway<arma::mat, arma::mat>*,
     Join<arma::mat, arma::mat>*,
     LayerNorm<arma::mat, arma::mat>*,
     LeakyReLU<arma::mat, arma::mat>*,
@@ -256,6 +261,7 @@ using LayerTypes = boost::variant<
     MultiplyConstant<arma::mat, arma::mat>*,
     MultiplyMerge<arma::mat, arma::mat>*,
     NegativeLogLikelihood<arma::mat, arma::mat>*,
+    NoisyLinear<arma::mat, arma::mat>*,
     Padding<arma::mat, arma::mat>*,
     PReLU<arma::mat, arma::mat>*,
     Softmax<arma::mat, arma::mat>*,