+++ HISTORY.md
@@ -1,5 +1,8 @@
 ### mlpack ?.?.?
 ###### ????-??-??
+  * Added Dueling DQN to q_networks, Noisy linear layer to ann/layer
+    and Empty loss to ann/loss_functions (#2414).
+
   * Storing and adding accessor method for action in q_learning (#2413).
 
   * Added accessor methods for ANN layers (#2321).
@@ -17,8 +20,15 @@
 
   * Add Load and Save of Sparse Matrix (#2344).
 
+  * Add Intersection over Union (IoU) metric for bounding boxes (#2402).
+
+  * Add Non Maximal Supression (NMS) metric for bounding boxes (#2410).
+
   * Fix `no_intercept` and probability computation for linear SVM bindings
     (#2419).
+
+  * Fix incorrect neighbors for `k > 1` searches in `approx_kfn` binding, for
+    the `QDAFN` algorithm (#2448).
 
 ### mlpack 3.3.1
 ###### 2020-04-29